+++ ETS2LA/Controls/picker.py
@@ -1,11 +1,9 @@
+from ETS2LA.Handlers.utils.key_mappings import key_to_str
+from pynput import keyboard as pynput_keyboard
 from ETS2LA.Controls import ControlEvent
 from ETS2LA.UI import SendPopup
 import multiprocessing
-import queue
 import time
-import os
-from pynput import keyboard as pynput_keyboard
-from ETS2LA.Handlers.key_mappings import key_to_str
 
 new_guid = ""
 new_key = ""
