+++ requirements.txt
@@ -27,7 +27,6 @@ gitpython
 dearpygui
 bettercam
 keyboard
-pynput
 requests
 ipython
 pywin32
@@ -38,6 +37,7 @@ pillow
 psutil
 pygame
 sv_ttk
+pynput
 pydub
 mouse
 scipy
